Add a button for confirmation for more accessibility
