+++ src/components/swaggerPlugins/VersionReducers.ts
@@ -6,5 +6,8 @@ export const VersionReducers = {
     API_VERSION_SET: (state: any, action: any) => {
       return state.set('apiVersion', action.payload);
     },
+    API_VERSION_UPDATED: (state: any, action: any) => {
+      return state.set('apiVersion', action.payload);
+    },
   },
 };
